Belt sanding pass fd_vm_syscalls
* Started as converting FD_VM_MEM_MAP_ERR codes to common integer error
  codes consistent with the rest of the APIs.
* Getting rid of all the spurious silent conversions between int and ulong
  required updating all the syscall returns.  Which then turned into
  a belt sanding pass on the syscalls.  Notable changes include:
  - returns a int for an error code
  - consistent naming an enumeration of the syscall arguments
  - eliminated the unused redundant syscall function pointer type
  - got rid of a bunch of int / ulong casting
  - got rid of a bunch of unnecessary function pointer casting
  - consume_compute API cleaned up, optimized, proper error hanlding and documented (similarly for mem_op_consume)
  - bunch of linting code in fd_vm_syscall (plenty more to do), including annotations of likely locations of latent (or
    not so latent) issues
* Some very minor changes to runtime and ballet/sbpf got caught up in
  this (actually fixed a silent cast in runtime)

Note that this has uncovered some worrisome bugs in the vm.  When
starting from a clean build on head of tree, I had to disable a bunch of
failing tests in test_vm_interp and test_vm_syscalls to begin work.

With this pass, though nothing here should have affected had an
functional impact (it is running about 2% faster from cleanups alone),
now all but 1 test in test_vm_syscall are passing.  But that 1 test
(log_sol_data) now seg faults.  Except when all the unrelated tests are
disabled.  Then the test fails in a different way.  All this and various
diagnostics collected to troubleshoot there are pre-existing latent
memory corruption issues (maybe in the test itself, maybe in the VM).

Unit test coverage here is pretty thin currently.  So, the VM will
almost certainly need a lot of security auditing and what not
after the belt sanding.

Plenty more belt sanding to do (especially at a high level, e.g. should
break fd_vm_syscall.c into a bunch of files to make auditing and testing
all this or manageable).

+++ src/flamenco/runtime/program/fd_native_cpi.c
@@ -84,11 +84,11 @@ fd_native_cpi_execute_system_program_instruction( fd_exec_instr_ctx_t * ctx,
 
   instr_info->data = buf;
   instr_info->data_sz = sizeof(buf);
-  ulong exec_err = fd_vm_prepare_instruction( ctx->instr, instr_info, ctx, instruction_accounts,
+  int exec_err = fd_vm_prepare_instruction( ctx->instr, instr_info, ctx, instruction_accounts,
                                               &instruction_accounts_cnt, signers, signers_cnt );
   if( exec_err != FD_EXECUTOR_INSTR_SUCCESS ) {
     FD_LOG_WARNING(("Preparing instruction failed"));
-    return (int)exec_err;
+    return exec_err;
   }
 
   return fd_execute_instr( ctx->txn_ctx, instr_info );
